Fix all remaining stylistic lint errors, more organization in preparation for refactor
