+++ src/st7701s_spi/mod.rs
@@ -2,3 +2,4 @@ pub mod commands;
 pub mod panel;
 pub mod sequences;
 pub mod spi;
+pub mod parameters;
