+++ src/st7701s_spi/sequences/sequence_tdo.rs
@@ -2,29 +2,26 @@ use std::{thread, time};
 
 use crate::st7701s_spi::{
     commands::{
-        BK0Command2, BK1Command2, BitsPerPixel, ColorOrder, Command, Command2Selection,
-        CommandsGeneral, EndPixelFormat, GammaOPBias, Inversion, LEDPolarity, PWMPolarity,
-        PixelPinout, ScanDirection, SourceOPInput, SourceOPOutput, VoltageAVCL, VoltageAVDD,
-    },
-    panel::Mode,
-    spi::HalfDuplexSPI,
+        BK0Command2, BK1Command2,  Command, Command2Selection,
+        CommandsGeneral,
+    }, panel::Mode, parameters::{BitsPerPixel, ColorOrder, EndPixelFormat, GammaOPBias, Inversion, LEDPolarity, PWMPolarity, PixelPinout, ScanDirection, SourceOPInput, SourceOPOutput, VoltageAVCL, VoltageAVDD}, spi::HalfDuplexSPI
 };
 
 pub fn init(display: &mut HalfDuplexSPI, mode: Mode) {
-    let mut CMD2: Command2Selection = Command2Selection::Disabled;
+    let mut cmd2: Command2Selection = Command2Selection::Disabled;
 
     // Set Command2 for BK0
     display.write_command(CommandsGeneral::set_command_2(Command2Selection::BK0));
-    CMD2 = Command2Selection::BK0;
+    cmd2 = Command2Selection::BK0;
 
-    display.write_command(BK0Command2::display_line_setting(&CMD2, 0x3B, 0x00, 0x00));
+    display.write_command(BK0Command2::display_line_setting(&cmd2, 0x3B, 0x00, 0x00));
     // Note: This will be off by one from the TDO spec:
     // SPI_WriteComm(0xC1); // PORCTRL
     // 0x0B); // V,
     // 0x02); // V,
-    display.write_command(BK0Command2::porch_control(&CMD2, &mode));
+    display.write_command(BK0Command2::porch_control(&cmd2, &mode));
     display.write_command(BK0Command2::inversion_select(
-        &CMD2,
+        &cmd2,
         Inversion::OneDot,
         0x02,
     ));
@@ -33,21 +30,21 @@ pub fn init(display: &mut HalfDuplexSPI, mode: Mode) {
         parameters: vec![0x10],
     }));
     display.write_command(BK0Command2::color_control(
-        &CMD2,
+        &cmd2,
         PWMPolarity::Low,
         LEDPolarity::Low,
         PixelPinout::Condensed,
         EndPixelFormat::SelfMSB,
     ));
     display.write_command(BK0Command2::positive_gamma_control(
-        &CMD2,
+        &cmd2,
         &[
             0x02, 0x13, 0x1B, 0x0D, 0x10, 0x05, 0x08, 0x07, 0x07, 0x24, 0x04, 0x11, 0x0E, 0x2C,
             0x33, 0x1D,
         ],
     ));
     display.write_command(BK0Command2::negative_gamma_control(
-        &CMD2,
+        &cmd2,
         &[
             0xB1, 0x05, 0x13, 0x1B, 0x0D, 0x11, 0x05, 0x08, 0x07, 0x07, 0x24, 0x04, 0x11, 0x0E,
             0x2C, 0x33, 0x1D,
@@ -56,27 +53,27 @@ pub fn init(display: &mut HalfDuplexSPI, mode: Mode) {
 
     // Set Command2 for BK1
     display.write_command(CommandsGeneral::set_command_2(Command2Selection::BK1));
-    CMD2 = Command2Selection::BK1;
+    cmd2 = Command2Selection::BK1;
 
-    display.write_command(BK1Command2::set_vop_amplitude(&CMD2, 0x5d));
-    display.write_command(BK1Command2::set_vcom_amplitude(&CMD2, 0x43));
-    display.write_command(BK1Command2::set_vgh_voltage(&CMD2, 0x81));
-    display.write_command(BK1Command2::test_command_setting(&CMD2));
-    display.write_command(BK1Command2::set_vgl_voltage(&CMD2, 0x43));
+    display.write_command(BK1Command2::set_vop_amplitude(&cmd2, 0x5d));
+    display.write_command(BK1Command2::set_vcom_amplitude(&cmd2, 0x43));
+    display.write_command(BK1Command2::set_vgh_voltage(&cmd2, 0x81));
+    display.write_command(BK1Command2::test_command_setting(&cmd2));
+    display.write_command(BK1Command2::set_vgl_voltage(&cmd2, 0x43));
     display.write_command(BK1Command2::power_control_one(
-        &CMD2,
+        &cmd2,
         GammaOPBias::Middle,
         SourceOPInput::Min,
         SourceOPOutput::Min,
     ));
 
     display.write_command(BK1Command2::power_control_two(
-        &CMD2,
+        &cmd2,
         VoltageAVDD::Pos6_6,
         VoltageAVCL::Neg4_4,
     ));
-    display.write_command(BK1Command2::set_pre_drive_timing_one(&CMD2, 0x08));
-    display.write_command(BK1Command2::set_pre_drive_timing_two(&CMD2, 0x08));
+    display.write_command(BK1Command2::set_pre_drive_timing_one(&cmd2, 0x08));
+    display.write_command(BK1Command2::set_pre_drive_timing_two(&cmd2, 0x08));
     display.write_command(Ok(Command {
         address: 0xD0,
         parameters: vec![0x88],
@@ -145,7 +142,7 @@ pub fn init(display: &mut HalfDuplexSPI, mode: Mode) {
 
     // Set Command2 for BK3
     display.write_command(CommandsGeneral::set_command_2(Command2Selection::BK1));
-    CMD2 = Command2Selection::BK1;
+    cmd2 = Command2Selection::BK1;
     display.write_command(Ok(Command {
         address: 0xEF,
         parameters: vec![0x08],
@@ -153,7 +150,7 @@ pub fn init(display: &mut HalfDuplexSPI, mode: Mode) {
 
     // COMMAND2 DISABLE
     display.write_command(CommandsGeneral::set_command_2(Command2Selection::Disabled));
-    CMD2 = Command2Selection::Disabled;
+    cmd2 = Command2Selection::Disabled;
 
     display.write_command(CommandsGeneral::sleep_mode_off());
     thread::sleep(time::Duration::from_millis(120));
